adding MatchList api method

+++ summoner.go
@@ -1,9 +1,9 @@
 package main
 
 type Summoner struct {
-	Id            int
+	Id            int64
 	Name          string
 	ProfileIconId int
-	SummonerLevel int
-	RevisionDate  int
+	SummonerLevel int64
+	RevisionDate  int64
 }
